requirements.txt
sqlalchemy
python-dotenv
pytest
psycopg2-binary
